requirements.txt
coverage==7.4.0  # for testing
django==4.2.13
gunicorn==22.0.0
psycopg2-binary==2.9.9
pytest-django==4.7.0  # for testing
pytest==7.4.4  # for testing
wait-for-it==2.2.2
django-mptt==0.14.0